move DetailModal to components/legacy
