+++ src/components/legacy/DetailModal/DetailModal.js
