+++ src/components/legacy/DetailModal/package.json
